fixed some gcc warnings, and removed non-standard code

+++ trunk/tools/base64.c
@@ -80,7 +80,7 @@ void *NewBase64Decode(
 					  size_t length,
 					  size_t *outputLength)
 {
-	if (length == -1)
+	if ((int) length == -1)
 	{
 		length = strlen(inputBuffer);
 	}
